# openapi.yaml
openapi: 3.0.3
info:
  title: ManadaBook API
  description: API completa para la red social de mascotas más grande del mundo
  version: 1.0.0
  contact:
    name: ManadaBook Support
    email: support@manadabook.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.manadabook.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
        avatar_url:
          type: string
          format: uri
        role:
          type: string
          enum: [user, admin, moderator]
        created_at:
          type: string
          format: date-time
      required: [id, name, email]
...
    Snippet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 500
        video_url:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
        duration:
          type: integer
          minimum: 0
        views:
          type: integer
          minimum: 0
        likes:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
      required: [id, author_id, title, video_url]
    Ad:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 500
        target:
          type: object
          properties:
            location:
              type: string
            species:
              type: array
              items:
                type: string
            interests:
              type: array
              items:
                type: string
        budget:
          type: number
          minimum: 0
        bid_type:
          type: string
          enum: [CPM, CPC, CPI]
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        creative:
          type: array
          items:
            type: string
            format: uri
        status:
          type: string
          enum: [pending, approved, active, paused, completed]
        impressions:
          type: integer
          minimum: 0
        clicks:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
      required: [id, owner_id, title, budget, bid_type]
...
    AnalyticsEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        session_id:
          type: string
        event_type:
          type: string
          enum: [page_view, post_create, post_like, snippet_view, ad_click, purchase, join_circle]
        event_category:
          type: string
          enum: [navigation, content, video, monetization, social]
        event_data:
          type: object
        page_url:
          type: string
          format: uri
        device_type:
          type: string
          enum: [desktop, mobile, tablet]
        created_at:
          type: string
          format: date-time
      required: [id, event_type, event_category]
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Registrar nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                phone:
                  type: string
              required: [name, email, password]
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /snippets:
    get:
      tags:
        - Snippets
      summary: Listar videos cortos
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 10
      responses:
        '200':
          description: Lista de videos
          content:
            application/json:
              schema:
                type: object
                properties:
                  snippets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Snippet'
                  total:
                    type: integer
                  page:
                    type: integer
                  hasMore:
                    type: boolean

    post:
      tags:
        - Snippets
      summary: Subir nuevo video
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
                description:
                  type: string
                  maxLength: 500
                video:
                  type: string
                  format: binary
      responses:
        '201':
          description: Video subido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
...
  /ads:
    get:
      tags:
        - Ads
      summary: Listar campañas publicitarias
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, active, paused, completed]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de campañas
          content:
            application/json:
              schema:
                type: object
                properties:
                  ads:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ad'
                  total:
                    type: integer
                  page:
                    type: integer
                  hasMore:
                    type: boolean

    post:
      tags:
        - Ads
      summary: Crear nueva campaña
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ad'
      responses:
        '201':
          description: Campaña creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ad'
...
  /analytics/events:
    get:
      tags:
        - Analytics
      summary: Obtener eventos de analytics
      parameters:
        - name: event_type
          in: query
          schema:
            type: string
        - name: event_category
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de eventos
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnalyticsEvent'
                  total:
                    type: integer
                  page:
                    type: integer
                  hasMore:
                    type: boolean

    post:
      tags:
        - Analytics
      summary: Trackear evento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event_type:
                  type: string
                event_category:
                  type: string
                event_data:
                  type: object
                page_url:
                  type: string
                  format: uri
                device_type:
                  type: string
                  enum: [desktop, mobile, tablet]
              required: [event_type, event_category]
      responses:
        '201':
          description: Evento trackeado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsEvent'
...